+++ config-auth/authmiddleware/secret.yaml
@@ -15,4 +15,4 @@ data:
   # 3. Production deployments use the Helm chart which generates random keys via:
   #    jwt-signing-key-{{ now | unixEpoch }}: {{ randBytes 48 | quote }}
   # 4. JWTs signed with this key are only valid within the local Kind cluster
-  jwt-signing-key-1700000000: "YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXoxMjM0NTY3ODkwQUJDREVGR0hJSktMTU5PUFFSU1RVVg=="
+  jwt-signing-key-1: "YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXoxMjM0NTY3ODkwQUJDREVGR0hJSktMTU5PUFFSU1RVVg=="
