+++ guided-charts/aws-hyperpod/templates/authmiddleware/deployment.yaml
@@ -69,7 +69,7 @@ spec:
           - name: PORT
             value: "8080"
           - name: NAMESPACE
-            value: "{{ .Values.namespace }}"
+            value: "{{ .Release.Namespace }}"
           - name: READ_TIMEOUT
             value: "{{ .Values.clusterWebUI.auth.readTimeout }}"
           - name: WRITE_TIMEOUT
